@@ -21,18 +21,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^17.3.8",
+    "@angular-devkit/build-angular": "^18.2.12",
     "@angular-eslint/schematics": "^17.5.1",
-    "@angular/animations": "^17.3.10",
-    "@angular/cli": "^17.3.8",
-    "@angular/common": "^17.3.10",
-    "@angular/compiler": "^17.3.10",
-    "@angular/compiler-cli": "^17.3.10",
-    "@angular/core": "^17.3.10",
-    "@angular/forms": "^17.3.10",
-    "@angular/platform-browser": "^17.3.10",
-    "@angular/platform-browser-dynamic": "^17.3.10",
-    "@angular/router": "^17.3.10",
+    "@angular/animations": "^18.2.13",
+    "@angular/cli": "^18.2.12",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/compiler-cli": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
     "@auth0/angular-jwt": "^5.1.0",
     "@rxweb/reactive-form-validators": "^2.1.6",
     "@sentry/angular": "^8.18.0",
@@ -50,7 +50,7 @@
     "ts-cacheable": "^1.0.8",
     "tslib": "^2.3.0",
     "typescript": "~5.4.5",
-    "zone.js": "~0.14.6"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
     "@angular-eslint/builder": "14.1.2",
