@@ -1,5 +1,5 @@
-import { HttpClientModule } from '@angular/common/http';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormBuilder } from '@angular/forms';
 import { By } from '@angular/platform-browser';
@@ -40,17 +40,17 @@ xdescribe('BambooHrComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ BambooHrComponent ],
-      imports: [
-        HttpClientModule, HttpClientTestingModule
-      ],
-      providers: [
+    declarations: [BambooHrComponent],
+    imports: [],
+    providers: [
         FormBuilder,
         MessageService,
         { provide: OrgService, useValue: service1 },
-        { provide: BambooHrService, useValue: service2 }
-      ]
-    })
+        { provide: BambooHrService, useValue: service2 },
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting()
+    ]
+})
     .compileComponents();
 
     fixture = TestBed.createComponent(BambooHrComponent);
