@@ -1,11 +1,12 @@
 import { TestBed, getTestBed } from '@angular/core/testing';
 
 import { QbdMappingService } from './qbd-mapping.service';
-import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
+import { HttpTestingController, provideHttpClientTesting } from '@angular/common/http/testing';
 import { environment } from 'src/environments/environment';
 import { MappingState } from 'src/app/core/models/enum/enum.model';
 import { QBDMapping, QBDMappingPost, QBDMappingResponse, QBDMappingStats } from 'src/app/core/models/qbd/db/qbd-mapping.model';
 import { QBDExportSettingResponse, QBDExportSettingResponse2 } from 'src/app/integrations/qbd/qbd-shared/qbd-export-setting/qbd-export-setting.fixture';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 xdescribe('QbdMappingService', () => {
   let service: QbdMappingService;
@@ -17,9 +18,9 @@ xdescribe('QbdMappingService', () => {
   beforeEach(() => {
     localStorage.setItem('workspaceId', '4');
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
-      providers: [QbdMappingService]
-    });
+    imports: [],
+    providers: [QbdMappingService, provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+});
     injector = getTestBed();
     service = TestBed.inject(QbdMappingService);
     httpMock = injector.inject(HttpTestingController);
