@@ -1,16 +1,17 @@
 import { TestBed } from '@angular/core/testing';
 
 import { SiAdvancedSettingService } from './si-advanced-setting.service';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 xdescribe('SiAdvancedSettingService', () => {
   let service: SiAdvancedSettingService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
-      providers: [SiAdvancedSettingService]
-    });
+    imports: [],
+    providers: [SiAdvancedSettingService, provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+});
     service = TestBed.inject(SiAdvancedSettingService);
   });
 
