@@ -1,8 +1,9 @@
 import { TestBed, getTestBed } from '@angular/core/testing';
 
 import { SiExportSettingService } from './si-export-setting.service';
-import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
+import { HttpTestingController, provideHttpClientTesting } from '@angular/common/http/testing';
 import { environment } from 'src/environments/environment';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 xdescribe('SiExportSettingService', () => {
   let service: SiExportSettingService;
@@ -14,9 +15,9 @@ xdescribe('SiExportSettingService', () => {
   beforeEach(() => {
     localStorage.setItem('workspaceId', '1');
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
-      providers: [SiExportSettingService]
-    });
+    imports: [],
+    providers: [SiExportSettingService, provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+});
     injector = getTestBed();
     httpMock = injector.inject(HttpTestingController);
     service = TestBed.inject(SiExportSettingService);
