@@ -1,10 +1,11 @@
-import { HttpTestingController, HttpClientTestingModule } from '@angular/common/http/testing';
+import { HttpTestingController, provideHttpClientTesting } from '@angular/common/http/testing';
 import { getTestBed, TestBed } from '@angular/core/testing';
 import { environment } from 'src/environments/environment';
 import { OrgService } from '../org/org.service';
 import { travelperkMockData, connectTravelperkMockData, connectAwsS3MockData, travelperkConfigurationMockData } from './travelperk.fixture';
 
 import { TravelperkService } from './travelperk.service';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 xdescribe('TravelperkService', () => {
   let service: TravelperkService;
@@ -18,13 +19,13 @@ xdescribe('TravelperkService', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [
-        HttpClientTestingModule
-      ],
-      providers: [
-        { provide: OrgService, useValue: service1 }
-      ]
-    });
+    imports: [],
+    providers: [
+        { provide: OrgService, useValue: service1 },
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting()
+    ]
+});
     injector = getTestBed();
     service = TestBed.inject(TravelperkService);
     orgService = injector.inject(OrgService);
