@@ -1,7 +1,8 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ConfigurationStepHeaderComponent } from './configuration-step-header.component';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 xdescribe('ConfigurationStepHeaderComponent', () => {
   let component: ConfigurationStepHeaderComponent;
@@ -9,9 +10,10 @@ xdescribe('ConfigurationStepHeaderComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
-      declarations: [ ConfigurationStepHeaderComponent ]
-    })
+    declarations: [ConfigurationStepHeaderComponent],
+    imports: [],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+})
     .compileComponents();
 
     fixture = TestBed.createComponent(ConfigurationStepHeaderComponent);
